Extract brief/reference-image routes out of agent-channel.ts (Wave D batch 3)
agent-channel.ts 2948 -> 2877 lines. Moves BRIEF/REFERENCE_IMAGES into
a new agent-channel-brief.ts as registerAgentChannelBriefRoutes(app, deps),
same pattern as batches 1-2.

Neither route touches the composition-root state (channelState,
markBuildingFromChannel, rate-limit Maps) -- the cleanest remaining
cluster per a fresh scoping pass. AGENT_BUILD_RULES_DIGEST/briefLocales/
buildConstraints/DEFAULT_BUILD_ORIENTATION and dispatchAttempt move
with the cluster (their only callers in this file); seedPayload stays
imported in agent-channel.ts (still used by SEED/SOURCES routes).
