+++ eslint-rules/module-boundary-map.mjs
@@ -176,6 +176,7 @@ const FILE_BUCKET = {
   'agent-channel': 'agent-surface',
   'agent-channel-kit-files': 'agent-surface',
   'agent-channel-examples': 'agent-surface',
+  'agent-channel-brief': 'agent-surface',
   'mcp-server': 'agent-surface',
   'mcp-tool-support': 'agent-surface',
   'mcp-example-tools': 'agent-surface',
